Making Parse public, adding helper method to create default CliContext
Supporting changes to allow calling code to inspect and modify parsed context before execution
