+++ cmd/spiffe-aws-assume-role/cli/cli_context.go
@@ -12,3 +12,24 @@ type CliContext struct {
 	Logger            *logrus.Logger
 	Telemetry         *telemetry.Telemetry
 }
+
+func NewDefaultCliContext() (*CliContext, error) {
+	logger := logrus.New()
+	// Example log line:
+	// time="2021-03-01T16:28:51-08:00" level=error msg="failed to parse SPIFFE ID from 305d07ed-765e-4642-9bd3-4c74aa86f5ed: spiffeid: invalid scheme"
+	logger.SetFormatter(&logrus.TextFormatter{})
+
+	telemetry, err := telemetry.NullTelemetry()
+	if err != nil {
+		return nil, err
+	}
+
+	context := &CliContext{
+		JWTSourceProvider: credentials.StandardJWTSourceProvider,
+		STSProvider:       credentials.StandardSTSProvider,
+		Logger:            logger,
+		Telemetry:         telemetry,
+	}
+
+	return context, nil
+}
